refactor: Separate build and deploy model for retrieval
