+++ jobs/ml_jobs/retrieval_vector/build_dummy_model.py
@@ -1,17 +1,13 @@
-from datetime import datetime
-
 import numpy as np
 import typer
 
 from utils import (
     ENV_SHORT_NAME,
     create_items_table,
-    deploy_container,
     get_item_docs,
     get_items_metadata,
     get_user_docs,
     get_users_dummy_metadata,
-    save_experiment,
     save_model_type,
 )
 
@@ -51,27 +47,9 @@ def prepare_docs():
     )
 
 
-def main(
-    experiment_name: str = typer.Option(
-        None,
-        help="Name of the experiment",
-    ),
-    model_name: str = typer.Option(
-        None,
-        help="Name of the model",
-    ),
-) -> None:
-    yyyymmdd = datetime.now().strftime("%Y%m%d")
-    if model_name is None:
-        model_name = "default"
-    run_id = f"{model_name}_{ENV_SHORT_NAME}_v{yyyymmdd}"
-    serving_container = f"europe-west1-docker.pkg.dev/passculture-infra-prod/pass-culture-artifact-registry/data-gcp/retrieval-vector/{ENV_SHORT_NAME}/{experiment_name.replace('.', '_')}:{run_id}"
-
+def main() -> None:
     prepare_docs()
-    print("Deploy...")
     save_model_type(model_type=MODEL_TYPE)
-    deploy_container(serving_container, workers=2)
-    save_experiment(experiment_name, model_name, serving_container, run_id=run_id)
 
 
 if __name__ == "__main__":
